requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
eokulapi==0.0.3
feedgen==0.9.0
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
